feat, fix: Organize files, fix: run toast on ui Thread
