+++ feature/feature-util/src/main/java/org/gampiot/robok/feature/util/di/Modules.kt
@@ -10,9 +10,9 @@ import org.koin.core.module.dsl.singleOf
 import org.koin.android.ext.koin.androidContext
 import org.koin.dsl.module
 
+import org.gampiot.robok.feature.settings.compose.viewmodels.ConfigureRDKViewModel
 import org.gampiot.robok.feature.settings.compose.viewmodels.AppPreferencesViewModel
 import org.gampiot.robok.feature.settings.compose.repositories.AppPreferencesRepository
-import org.gampiot.robok.feature.settings.compose.screens.ui.rdkmanager.ConfigureRDKViewModel
 
 const val APP_PREFERENCES = "app_preferences"
